+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/RasterMappedTo3DTile.h
@@ -210,10 +210,7 @@ public:
    *
    * @param maximumScreenSpaceError The maximum screen space error that is used
    * for the current tile
-   * @param tileProvider The overlay tile provider to map to the tile. This may
-   * be a placeholder if the tile provider is not yet ready.
-   * @param placeholder The placeholder tile provider for this overlay. This is
-   * always a placeholder, even if the tile provider is already ready.
+   * @param activatedOverlay The activated overlay to map to the tile.
    * @param tile The tile to which to map the overlay.
    * @param missingProjections The list of projections for which there are not
    * yet any texture coordiantes. On return, the given overlay's Projection may