+++ CesiumRasterOverlays/include/CesiumRasterOverlays/RasterOverlay.h
@@ -191,9 +191,9 @@ public:
   /**
    * @brief Activates this overlay.
    *
-   * This method is called by a {@link RasterOverlayCollection} when an overlay
-   * is added to it. The returned {@link ActivatedRasterOverlay} is used by the
-   * collection to manage the overlay.
+   * This method is called by a {@link Cesium3DTilesSelection::RasterOverlayCollection}
+   * when an overlay is added to it. The returned {@link ActivatedRasterOverlay}
+   * is used by the collection to manage the overlay.
    *
    * @param externals The external interfaces for use by the raster overlay.
    * @param ellipsoid The {@link CesiumGeospatial::Ellipsoid}.